requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.4
colorama==0.4.4
dataclasses==0.6
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
typing_extensions==4.1.1
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
